requirements.txt
alembic==1.10.2
biopython @ file:///C:/ci_310/biopython_1642078264950/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
packaging==23.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
Werkzeug==2.2.3
wincertstore==0.2
WTForms==3.0.1
